Move getGroups method to secondary navigation
