+++ shell/components/NavigationSecondary.vue
@@ -1,9 +1,17 @@
 <script>
 import { mapState, mapGetters } from 'vuex';
-import { CATALOG } from '@shell/config/types';
+import { CATALOG, UI } from '@shell/config/types';
 import { HARVESTER_NAME as HARVESTER } from '@shell/config/features';
 import { getVersionInfo } from '@shell/utils/version';
 import Group from '@shell/components/nav/Group';
+import { replaceWith, clear, addObject, addObjects } from '@shell/utils/array';
+import { NAME as EXPLORER } from '@shell/config/product/explorer';
+import { sortBy } from '@shell/utils/sort';
+import { BASIC, FAVORITE, USED } from '@shell/store/type-map';
+import { NAME as NAVLINKS } from '@shell/config/product/navlinks';
+import { ucFirst } from '@shell/utils/string';
+import debounce from 'lodash/debounce';
+import isEqual from 'lodash/isEqual';
 
 export default {
   components: { Group },
@@ -11,6 +19,7 @@ export default {
     ...mapState(['managementReady', 'clusterReady']),
     ...mapGetters(['clusterId', 'isSingleProduct', 'isExplorer']),
     ...mapGetters({ locale: 'i18n/selectedLocaleLabel', availableLocales: 'i18n/availableLocales' }),
+    ...mapGetters('type-map', ['activeProducts']),
 
     showClusterTools() {
       return this.isExplorer &&
@@ -49,14 +58,255 @@ export default {
 
       return displayVersion;
     },
+
+    allNavLinks() {
+      if ( !this.clusterId || !this.$store.getters['cluster/schemaFor'](UI.NAV_LINK) ) {
+        return [];
+      }
+
+      return this.$store.getters['cluster/all'](UI.NAV_LINK);
+    },
   },
-  props: {
-    groups: {
-      type:    Array,
-      default: () => [],
-    }
+  watch: {
+    counts(a, b) {
+      if ( a !== b ) {
+        this.queueUpdate();
+      }
+    },
+
+    allSchemas(a, b) {
+      if ( a !== b ) {
+        this.queueUpdate();
+      }
+    },
+
+    allNavLinks(a, b) {
+      if ( a !== b ) {
+        this.queueUpdate();
+      }
+    },
+
+    favoriteTypes(a, b) {
+      if ( !isEqual(a, b) ) {
+        this.queueUpdate();
+      }
+    },
+
+    locale(a, b) {
+      if ( !isEqual(a, b) ) {
+        this.getGroups();
+      }
+    },
+
+    namespaceMode(a, b) {
+      if ( !isEqual(a, b) ) {
+        // Immediately update because you'll see it come in later
+        this.getGroups();
+      }
+    },
+
+    namespaces(a, b) {
+      if ( !isEqual(a, b) ) {
+        // Immediately update because you'll see it come in later
+        this.getGroups();
+      }
+    },
+
+    clusterReady(a, b) {
+      if ( !isEqual(a, b) ) {
+        // Immediately update because you'll see it come in later
+        this.getGroups();
+      }
+    },
+
+    product(a, b) {
+      if ( !isEqual(a, b) ) {
+        // Immediately update because you'll see it come in later
+        this.getGroups();
+      }
+    },
+
+    productId(a, b) {
+      if ( !isEqual(a, b) ) {
+        // Immediately update because you'll see it come in later
+        this.getGroups();
+      }
+    },
+
+  },
+  created() {
+    this.queueUpdate = debounce(this.getGroups, 500);
+
+    this.getGroups();
+  },
+  data() {
+    return {
+      groups:        [],
+      gettingGroups: false,
+    };
   },
   methods: {
+    /**
+     * Fetch navigation by creating groups from product schemas
+     */
+    getGroups() {
+      if ( this.gettingGroups ) {
+        return;
+      }
+      this.gettingGroups = true;
+
+      if ( !this.clusterReady ) {
+        clear(this.groups);
+        this.gettingGroups = false;
+
+        return;
+      }
+
+      const currentProduct = this.$store.getters['productId'];
+      let namespaces = null;
+
+      if ( !this.$store.getters['isAllNamespaces'] ) {
+        const namespacesObject = this.$store.getters['namespaces']();
+
+        namespaces = Object.keys(namespacesObject);
+      }
+
+      // Always show cluster-level types, regardless of the namespace filter
+      const namespaceMode = 'both';
+      const out = [];
+      const loadProducts = this.isExplorer ? [EXPLORER] : [];
+
+      const productMap = this.activeProducts.reduce((acc, p) => {
+        return { ...acc, [p.name]: p };
+      }, {});
+
+      if ( this.isExplorer ) {
+        for ( const product of this.activeProducts ) {
+          if ( product.inStore === 'cluster' ) {
+            addObject(loadProducts, product.name);
+          }
+        }
+      }
+
+      // This should already have come into the list from above, but in case it hasn't...
+      addObject(loadProducts, currentProduct);
+
+      this.getProductsGroups(out, loadProducts, namespaceMode, namespaces, productMap);
+      this.getExplorerGroups(out);
+
+      replaceWith(this.groups, ...sortBy(out, ['weight:desc', 'label']));
+      this.gettingGroups = false;
+    },
+
+    getProductsGroups(out, loadProducts, namespaceMode, namespaces, productMap) {
+      const clusterId = this.$store.getters['clusterId'];
+      const currentType = this.$route.params.resource || '';
+
+      for ( const productId of loadProducts ) {
+        const modes = [BASIC];
+
+        if ( productId === NAVLINKS ) {
+          // Navlinks produce their own top-level nav items so don't need to show it as a product.
+          continue;
+        }
+
+        if ( productId === EXPLORER ) {
+          modes.push(FAVORITE);
+          modes.push(USED);
+        }
+
+        for ( const mode of modes ) {
+          const types = this.$store.getters['type-map/allTypes'](productId, mode) || {};
+
+          const more = this.$store.getters['type-map/getTree'](productId, mode, types, clusterId, namespaceMode, namespaces, currentType);
+
+          if ( productId === EXPLORER || !this.isExplorer ) {
+            addObjects(out, more);
+          } else {
+            const root = more.find(x => x.name === 'root');
+            const other = more.filter(x => x.name !== 'root');
+
+            const group = {
+              name:     productId,
+              label:    this.$store.getters['i18n/withFallback'](`product.${ productId }`, null, ucFirst(productId)),
+              children: [...(root?.children || []), ...other],
+              weight:   productMap[productId]?.weight || 0,
+            };
+
+            addObject(out, group);
+          }
+        }
+      }
+    },
+
+    getExplorerGroups(out) {
+      if ( this.isExplorer ) {
+        const allNavLinks = this.allNavLinks;
+        const toAdd = [];
+        const haveGroup = {};
+
+        for ( const obj of allNavLinks ) {
+          if ( !obj.link ) {
+            continue;
+          }
+
+          const groupLabel = obj.spec.group;
+          const groupSlug = obj.normalizedGroup;
+
+          const entry = {
+            name:        `link-${ obj._key }`,
+            link:        obj.link,
+            target:      obj.actualTarget,
+            label:       obj.labelDisplay,
+            sideLabel:   obj.spec.sideLabel,
+            iconSrc:     obj.spec.iconSrc,
+            description: obj.spec.description,
+          };
+
+          // If there's a spec.group (groupLabel), all entries with that name go under one nav group
+          if ( groupSlug ) {
+            if ( haveGroup[groupSlug] ) {
+              continue;
+            }
+
+            haveGroup[groupSlug] = true;
+
+            toAdd.push({
+              name:     `navlink-group-${ groupSlug }`,
+              label:    groupLabel,
+              isRoot:   true,
+              // This is the item that actually shows up in the nav, since this outer group will be invisible
+              children: [
+                {
+                  name:  `navlink-child-${ groupSlug }`,
+                  label: groupLabel,
+                  route: {
+                    name:   'c-cluster-navlinks-group',
+                    params: {
+                      cluster: this.clusterId,
+                      group:   groupSlug,
+                    }
+                  },
+                }
+              ],
+              weight: -100,
+            });
+          } else {
+            toAdd.push({
+              name:     `navlink-${ entry.name }`,
+              label:    entry.label,
+              isRoot:   true,
+              // This is the item that actually shows up in the nav, since this outer group will be invisible
+              children: [entry],
+              weight:   -100,
+            });
+          }
+        }
+
+        addObjects(out, toAdd);
+      }
+    },
+
     collapseAll() {
       this.$refs.groups.forEach((grp) => {
         grp.isExpanded = false;
